+++ src/Application/VortexLoader/S57_TracksAndRoutesA.cs
@@ -49,8 +49,8 @@ namespace S100Framework.Applications
                 switch (fcSubtype) {
                     case 1: { // DWRTPT_DeepWaterRoutePart
 
-                            var orientationValue = current.ORIENT.HasValue && current.ORIENT.Value != -32767d ? current.ORIENT!.Value : default(double?);
-                            var depthValue = current.DRVAL1.HasValue && current.DRVAL1.Value != -32767d ? current.DRVAL1!.Value : default(double?);
+                            var orientationValue = current.ORIENT.HasValue && current.ORIENT.Value != -32767m ? current.ORIENT!.Value : default(decimal?);
+                            var depthValue = current.DRVAL1.HasValue && current.DRVAL1.Value != -32767m ? current.DRVAL1!.Value : default(decimal?);
                             var trafficFlow = EnumHelper.GetEnumValue(current.TRAFIC!.Value);
 
                             var instance = new DeepWaterRoutePart {
@@ -72,7 +72,7 @@ namespace S100Framework.Applications
 
 
                             if (current.ORIENT.HasValue) {
-                                instance.orientationValue = current.ORIENT.Value != -32767d ? current.ORIENT.Value : null;
+                                instance.orientationValue = current.ORIENT.Value != -32767m ? current.ORIENT.Value : null;
                             }
 
 
@@ -152,7 +152,7 @@ namespace S100Framework.Applications
                             var instance = new Fairway();
 
                             if (current.DRVAL1.HasValue) {
-                                instance.depthRangeMinimumValue = current.DRVAL1.Value != -32767d ? current.DRVAL1.Value : null;
+                                instance.depthRangeMinimumValue = current.DRVAL1.Value != -32767m ? current.DRVAL1.Value : null;
                             }
 
                             instance.featureName = GetFeatureName(current.OBJNAM, current.NOBJNM);
@@ -167,7 +167,7 @@ namespace S100Framework.Applications
                             // TODO: maximumPermittedDraught
 
                             if (current.ORIENT.HasValue) {
-                                instance.orientationValue = current.ORIENT.Value != -32767d ? current.ORIENT.Value : null;
+                                instance.orientationValue = current.ORIENT.Value != -32767m ? current.ORIENT.Value : null;
                             }
 
                             if (current.QUASOU != default) {
@@ -453,7 +453,7 @@ namespace S100Framework.Applications
 
                             //// TODO: maximumPermittedDraught
 
-                            //if (current.ORIENT.HasValue && current.ORIENT.Value != -32767d) {
+                            //if (current.ORIENT.HasValue && current.ORIENT.Value != -32767m) {
                             //    instance.orientationValue = current.ORIENT.Value;
                             //}
 
@@ -709,7 +709,7 @@ namespace S100Framework.Applications
                             }
 
                             if (current.DRVAL1.HasValue) {
-                                instance.depthRangeMinimumValue = current.DRVAL1.Value != -32767d ? current.DRVAL1.Value : null;
+                                instance.depthRangeMinimumValue = current.DRVAL1.Value != -32767m ? current.DRVAL1.Value : null;
                             }
 
                             DateHelper.TryGetFixedDateRange(current.DATSTA, current.DATEND, out var dateRange);
@@ -720,7 +720,7 @@ namespace S100Framework.Applications
                             // TODO: interoperabilityIdentifier
 
                             if (current.ORIENT.HasValue) {
-                                instance.orientationValue = current.ORIENT.Value != -32767d ? current.ORIENT.Value : null;
+                                instance.orientationValue = current.ORIENT.Value != -32767m ? current.ORIENT.Value : null;
                             }
 
                             if (current.QUASOU != default) {
